requirements.txt
streamlit
numpy
pandas
biopython
tensorflow
